requirements.txt
# HeAR - Health Acoustic Representations ML Pipeline
# Requirements for respiratory disease detection using HeAR model

# Core ML/DL Frameworks
tensorflow>=2.15.0
tensorflow-hub>=0.16.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
audioread>=3.0.1
pydub>=0.25.1

# Data Science & Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Hugging Face (for HeAR model)
huggingface-hub>=0.20.0
keras>=3.0.0

# Kaggle API for dataset downloads
kaggle>=1.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
tqdm>=4.65.0

# Jupyter Notebooks
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
notebook>=7.0.0

# Model Evaluation & Metrics
imbalanced-learn>=0.11.0

# Git for cloning repositories
gitpython>=3.1.40

# Utilities
python-dotenv>=1.0.0
pathlib2>=2.3.7
requests>=2.31.0
PyYAML>=6.0.1

# Optional: Google Cloud (for Vertex AI deployment)
# google-cloud-aiplatform>=1.38.0
# google-cloud-storage>=2.13.0
